Split up Mesh.ts
