+++ core/frontend/src/render/webgl/CachedGeometry.ts
@@ -23,7 +23,7 @@ import { GL } from "./GL";
 import { BufferHandle, BufferParameters, BuffersContainer, QBufferHandle2d, QBufferHandle3d } from "./AttributeBuffers";
 import { InstancedGeometry } from "./InstancedGeometry";
 import { MaterialInfo } from "./Material";
-import { MeshGeometry } from "./Mesh";
+import { MeshGeometry } from "./MeshGeometry";
 import { EdgeGeometry, SilhouetteEdgeGeometry } from "./EdgeGeometry";
 import { SurfaceGeometry } from "./SurfaceGeometry";
 import { PointCloudGeometry } from "./PointCloud";
