+++ core/frontend/src/render/webgl/EdgeGeometry.ts
@@ -6,7 +6,6 @@
  * @module WebGL
  */
 
-import { MeshData, MeshGeometry } from "./Mesh";
 import { assert, dispose } from "@itwin/core-bentley";
 import { RenderMode } from "@itwin/core-common";
 import { TesselatedPolyline } from "../primitives/PolylineParams";
@@ -22,6 +21,8 @@ import { RenderOrder, RenderPass } from "./RenderFlags";
 import { System } from "./System";
 import { Target } from "./Target";
 import { TechniqueId } from "./TechniqueId";
+import { MeshData } from "./MeshData";
+import { MeshGeometry } from "./MeshGeometry";
 
 /** @internal */
 export class EdgeGeometry extends MeshGeometry {
